12/02/2025

+++ KOP/KOP.Common/DTOs/AssessmentDTOs/AssessmentResultValueDTO.cs
@@ -2,7 +2,7 @@
 {
     public class AssessmentResultValueDto
     {
-        public double Value { get; set; }
+        public int Value { get; set; }
         public int AssessmentMatrixRow { get; set; }
     }
 }
